refactor: remove license headers from source files
# Motivation

The policy of adding license headers are not
consistent in this project and is error prone to
maintain. Other repositories in the same
organization also do not have license headers.

Until we decide to add license headers to all
source files, removing the ones we have ensures
that we at least have consistency.

It is also unclear if the license headers are
needed at all, and could potentially be avoided by
using a `NOTICE` file instead, according to
https://infra.apache.org/licensing-howto.html#step-by-step