refactor: separation of arguments and logic
# Motivation

To increase readability and testability,
separating the command line arguments from the
business logic is a good idea.

As a step in this direction, this commit creates
new functions for handling the arguments, even if
all argument handling is not done in these
functions just yet.

Follow up commits will further make this
separation clearer.