refactor!: remove hidden `aart` command
# Motivation

`aart` is a hidden command that is not in any way
advertised to the outside world, and does not
provide any useful functionality.

It is only basically code we refactor and
maintain, but never use. So it is better to simply
delete the code and remove the maintenance burden.

Also ran `go mod tidy` to remove the unused
dependencies.