refactor: consolidate workflows
# Motivation

To the workflows easier to work with, we should
use the `jobs` parameter more often. Especially
when the triggers are the same.

This makes it easier to inspect the various jobs
as they are now mostly in the same workflows.

In addition, there should be dependencies between
the various jobs, for example, so that the tests
run on the actual release. Or not to automatically
release if the `lint` job fails.

# Changes

* `release.yml` is renamed to `main.yml`
* The jobs of `test.yml`, and `lint.yml` have now
also moved to `main.yml`.
* The job `goreleaser` has now been renamed to
  `Build`, as it reflects what it does better
  (nothing is actually released, and `goreleaser`
  is just an implementation detail).

# Future work

`github-pages` could also potentially be moved to
`main.yml`, but because the triggers are different
it has been left out for now.