ci: add benchmark job to workflow
# Motivation

To avoid performance regressions, this commit adds
a benchmark job to the CI workflow.

The idea is that if a given PR decreases the
performance sufficiently, (e.g. 150% or more), the
CI will fail and the PR cannot be merged.

The benchmark results should be visible at
https://nlnwa.github.io/warchaeology/benchmarks/ubuntu-22.04
after this commit has been merged.

# Changes

Added a publish benchmark job in the
`publish_benchmarks.yml` that also runs the
benchmarks and publishes them to the github pages
branch (currently `gh-pages`).

# Quirks

We need to re-run the benchmark results since
getting artifacts from a separate workflow is not
trivial. Ideally we should just run the benchmarks
once.

Publishing benchmarks must live in a separate
workflow, since the permissions it needs to write
to `gh-pages` should not be exposed to external
users creating PRs targeting this repository.

An alternative would be to not run workflows when
receiving external PRs, but that would make the
developer experience for contributors worse.

# Future work

After this PR has been merged, doing the actual
comparison should be a priority. Currently the
benchmarks are only published to the `gh-pages`
branch, and not actually compared to anything.

Add `Windows` to the benchmark matrix.