ci: push benchmark results to dedicated branch
# Motivation

The current benchmarking system does not compare
the results with previous benchmarks, as we have
no place to store the results.

This commit ensures that this is no longer the
case by pushing the benchmark results to a
dedicated branch called `benchmark-results`.

# Changes

After pushing the benchmark results to `GitHub
pages`, a json is now created and automatically
committed to the `benchmark-results` branch.

# Rejected alternatives

Use `actions/cache` to store the results. The
reason why this was not done is that the entries
in the cache will expire after a set amount of
time, which is likely to happen. In such a
scenario there will be additional burdens on the
developers to manually update the cache.

# Quirks

The benchmark results had to be moved out of the
working directory because of conflicts with the
checkout actions.