ci: compare with previous benchmarks
# Motivation

The previous commits of
343fb0a3b0d48856cb1af9237c43d0601f728a63 and
8f539bf9177c182a8d600cea02076538fc95a048 added
creation and storage of benchmark results.

This commit now does the final stage which is to
compare the new benchmark results with the
previous results.

# Changes

Renamed the benchmark name to `Go Benchmark` so
that it matches the benchmark name uploaded by
`publish_benchmarks.yml`, otherwise the comparison
will never work.

Use download previous benchmarks from the
dedicated `benchmark-results` branch.