test: disable failing tests on windows
# Motivation

To be able to support Windows to a higher degree,
testing must be enabled for Windows.

Unfortunately, a few of the tests are failing when
run on Windows. This commit disables these tests
when on Windows.

# Changes

When a test that fails on Windows is run, it is
skipped.

# Trade-offs

Ideally the test should always be run and expect
to fail instead. This was not trivial to implement
without refactoring the tests.