test: disable failing tests on MacOS
# Motivation

Similar to commit
d6595d1782aa4cbe97aa0246329853ee02d055cc

To be able to support MacOS to a higher degree,
testing must be enabled for MacOS.

Unfortunately, a few of the tests are failing when
run on MacOS. This commit disables these tests
when on MacOS.

# Changes

When a test that fails on MacOS is run, it is
skipped.

# Trade-offs

Ideally the test should always be run and expect
to fail instead. This was not trivial to implement
without refactoring the tests.

skip failing test on macOS