fix: add error handling to `workerPool`
# Motivation

The functions submitted to `workerPool` might
fail, and the errors they return should be
properly propagated.

# Changes

To propagate the errors to the main goroutine, a
channel is used to transfer the errors. When
closing the `workerPool` any errors transferred
will be evaluated and if there is any none-nil
errors, the first one will be returned.

# Quirks

This commit also disables the sub-linter `gocyclo`
for a function in `filewalker.go` as it is the
refactoring triggered by this commit increase the
complexity.

This complexity is only intermediate and will be
removed upon further refactoring.

The original linter error was the following:
```
lint: internal/filewalker/filewalker.go#L192
cyclomatic complexity 31 of func `(*fileWalker).Walk`
is high (> 30) (gocyclo)
```