test: add simplest test for `ls`
# Motivation

Many top level commands are missing tests. This is
a first step to add a test for `ls` with proper
test data.

# Changes

This commit adds a very simple test for `ls` that
checks that it does not crash when reading a file.

# Future work

Add more checks to the test so that it is even
better at avoiding regressions.

# Acknowledgements

Thanks to Ilya Kreymer for providing the test data
through issue
https://github.com/nlnwa/warchaeology/issues/99