ci: enable windows benchmarks
This commit is not quite a revert of commit
a91ef1303089616d253cd508783f429e3d2a85d9, but it
is close. The main difference is that this commit
handles the creation of the directories and
verification of the `push` status in an OS
agnostic way.

To avoid race conditions, each of the benchmark
jobs are now running one at a time, otherwise
there might be upload collision.