fix: silence benchmark output
Since commit
d8d316a6707a8f3e2234e5c483250ab7b99943de the
benchmark results of `BenchmarkReadFileWithError`
are no longer present.
`BenchmarkReadFileWithError` has not been deleted,
and it seems like some of the output printed to
the benchmark result file causes
`github-action-benchmark` to fail finding this
result.

The reason why commit
d8d316a6707a8f3e2234e5c483250ab7b99943de triggered
this error is because it added the required `lfs`
checkout so that the test-data was cloned
correctly.

As a workaround, this commit silences the output
of the benchmark `BenchmarkReadFileWithError`.