feat: WARC record iterator
The gowarc.WarcFileReader does not provide a simple way to get the size of the next
WARC record.

This commit introduces an iterator abstraction for iterating over WARC records. In
addition to returning the size of each record the iterator encapsulates the filtering
of records, limiting the number of records and selecting the nth record.