fix: avoid infinite loop if missing digest during deduplication
This commit fixes an infinite loop that happened when a record was missing
digest headers and the repair flag was not enabled.

Enabling the repair flag will generate missing digest headers when reading a
record. The generated digest headers can then be used to decide if a revisit
record shall be written or not.

If digest headers are generated and a revisit record is written, the revisit
record's payload digest header will be the payload digest of the referenced
record, and the revisit record's block digest header will be the digest of
the revisit record's block.